feature: Add links to character/contact/crew generator from world, with best-effort attempt att matching species and faction to world.
