+++ TravellerTools/Grauenwolf.TravellerTools.Web/Controls/WorldPicker.razor
@@ -135,6 +135,9 @@
                     <a href="@PlanetUrl("store")"><span class="oi oi-cart" aria-hidden="true" /> Store</a>
                     <a href="@PlanetUrl("animals")"><span class="oi oi-bug" aria-hidden="true" /> Animals</a>
                     <a href="@PlanetUrl("encounters")"><span class="oi oi-puzzle-piece" aria-hidden="true" /> Encounters</a>
+                    <a href="@PlanetUrl("character")"><span class="oi oi-person" aria-hidden="true" /> Character</a>
+                    <a href="@PlanetUrl("contacts")"><span class="oi oi-people" aria-hidden="true" /> Contacts</a>
+                    <a href="@PlanetUrl("crew")"><span class="oi oi-briefcase" aria-hidden="true" /> Crew</a>
                 </td>
             </tr>
         </table>
