+++ TravellerTools/Grauenwolf.TravellerTools.Web/Pages/CharacterPage.razor
@@ -1,4 +1,6 @@
 ﻿@page "/character"
+@page "/world/{milieuCode}/{sectorHex}/{planetHex}/character"
+@page "/uwp/{uwp}/character"
 @inherits NullablePageBase<CharacterOptions>
 @{
     PageTitle = "Character Generator";
