+++ TravellerTools/Grauenwolf.TravellerTools.Web/Pages/ContactsPage.razor
@@ -1,4 +1,6 @@
 ﻿@page "/contacts"
+@page "/world/{milieuCode}/{sectorHex}/{planetHex}/contacts"
+@page "/uwp/{uwp}/contacts"
 @using Microsoft.AspNetCore.WebUtilities
 @inherits NullablePageBase<ContactOptions>
 @{
