+++ TravellerTools/Grauenwolf.TravellerTools.Web/Pages/ContactsPage.razor.cs
@@ -1,4 +1,5 @@
 ﻿using Grauenwolf.TravellerTools.Characters;
+using Grauenwolf.TravellerTools.Maps;
 using Grauenwolf.TravellerTools.Names;
 using Grauenwolf.TravellerTools.Web.Data;
 using Microsoft.AspNetCore.Components;
@@ -9,9 +10,22 @@ namespace Grauenwolf.TravellerTools.Web.Pages;
 
 partial class ContactsPage
 {
+    [Parameter]
+    public string? MilieuCode { get; set; }
+
+    [Parameter]
+    public string? PlanetHex { get; set; }
+
+    [Parameter]
+    public string? SectorHex { get; set; }
+
+    [Parameter]
+    public string? Uwp { get; set; }
+
     protected ContactsModel? ContactsModel { get; set; }
     [Inject] CharacterBuilder CharacterBuilder { get; set; } = null!;
     [Inject] NameGenerator NameGenerator { get; set; } = null!;
+    [Inject] TravellerMapServiceLocator TravellerMapServiceLocator { get; set; } = null!;
 
     protected void GenerateContacts()
     {
@@ -49,6 +63,20 @@ partial class ContactsPage
     protected override void Initialized()
     {
         Model = new ContactOptions(CharacterBuilder);
+    }
+
+    protected override async Task ParametersSetAsync()
+    {
+        if (Model == null)
+            return;
+
+        Model.SpeciesOrFaction = await SpeciesOrFactionSelection.ResolveForWorldAsync(
+            Model.SpeciesOrFaction,
+            MilieuCode,
+            SectorHex,
+            PlanetHex,
+            TravellerMapServiceLocator,
+            Model.SpeciesAndFactionsList).ConfigureAwait(false);
     }
 
     protected string Permalink()
