+++ TravellerTools/Grauenwolf.TravellerTools.Web/Pages/CrewPage.razor
@@ -1,4 +1,6 @@
 ﻿@page "/crew"
+@page "/world/{milieuCode}/{sectorHex}/{planetHex}/crew"
+@page "/uwp/{uwp}/crew"
 @using Microsoft.AspNetCore.WebUtilities
 @inherits NullablePageBase<CrewOptions>
 @{
